+++ SailBoat-Rental/SailBoat-Rental/Service/Calculator/SmallBoatCalculator.cs
@@ -6,7 +6,7 @@
 
     public class SmallBoatCalculator : AbstractCalculator
     {
-        public override float Calculate(CalculatorArgs args)
+        public override double Calculate(CalculatorArgs args)
         {
             if (args is null)
             {
@@ -14,7 +14,7 @@
             }
                 
             
-            return args.BasicFee * 1.2f + args.HourlyRate * 1.3f * base.getNumberOfHours(args);
+            return args.BasicFee * 1.2d + args.HourlyRate * 1.3d * base.getNumberOfHours(args);
         }
     }
 }
